tools/seistech_utils/requirements.txt
Flask==1.1.2
GitPython==3.1.13
itsdangerous==2.0.1
numpy==1.20.1
pandas==1.2.2
PyYAML==5.4.1
requests==2.25.1
slack-sdk==3.7.0